pyproject.toml
[project]
name = "scan-to-text"
version = "0.1.0"
description = "OCR processing of scanned letters via Mistral OCR"
authors = [
    {name = "Sylvain Jalbert"}
]
requires-python = ">=3.10"
dependencies = [
    "mistralai (>=2.0.4,<3.0.0)",
    "openai (>=1.0.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
